# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 刷题 App API
  version: 0.1.0
  description: |
    首版接口契约（阶段 1：可靠刷题闭环 + 学习档案 + 管理端内容维护）。
    - 前缀 /api/v1；JSON 字段 camelCase，枚举值小写 snake_case。
    - 时间为 UTC RFC 3339；列表使用游标分页 ?cursor=&limit=（默认 20，最大 100）。
    - 认证：HttpOnly Cookie（session）；写请求校验 Origin。
    - 错误统一结构：{"error": {"code", "message", "requestId"}}；前端分支只依赖稳定 code。
    - 核心边界：提交前接口永不返回答案、解析或判分字段；answer_keys 仅判分与管理端可读。
servers:
  - url: /api/v1
...
paths:
...
  /admin/questions:
    get:
      tags: [admin]
      summary: 题目列表（含答案覆盖筛选；游标分页）
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [draft, in_review, published, retired] } }
        - { name: levelId, in: query, schema: { type: string } }
        - { name: subjectId, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: hasAnswer, in: query, schema: { type: string, enum: [yes, no] } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  questions:
                    type: array
                    items: { $ref: '#/components/schemas/QuestionAdmin' }
                  nextCursor: { type: string }
    post:
      tags: [admin]
      summary: 新建题目（第一版；可携带材料与标准答案）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuestionInput' }
      responses:
        '201': { description: OK, content: { application/json: { schema: { type: object, properties: { question: { $ref: '#/components/schemas/QuestionAdmin' } } } } } }
  /admin/questions/{id}:
    get:
      tags: [admin]
      summary: 题目详情（管理端可读标准答案）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { question: { $ref: '#/components/schemas/QuestionAdmin' } } } } } }
    patch:
      tags: [admin]
      summary: 编辑题目（总是创建新版本；已发布版本继续对外服务）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuestionInput' }
      responses:
        '200': { description: OK }
        '409': { description: question_retired }
  /admin/questions/{id}/submit-review:
    post:
      tags: [admin]
      summary: 草稿提交审核
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /admin/questions/{id}/publish:
    post:
      tags: [admin]
      summary: 发布当前版本（事务内校验；写入审计）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '409': { description: question_invalid / already_published / question_retired }
  /admin/questions/{id}/retire:
    post:
      tags: [admin]
      summary: 下架题目（历史记录保留题目快照）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /admin/knowledge-points:
    get:
      tags: [admin]
      summary: 知识点管理列表
      parameters: [{ name: levelId, in: query, schema: { type: string } }]
      responses: { '200': { description: OK } }
    post:
      tags: [admin]
      summary: 新建知识点（默认草稿）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, levelId, subjectId]
              properties:
                name: { type: string }
                levelId: { type: string }
                subjectId: { type: string }
                parentId: { type: string, nullable: true }
                description: { type: string }
                commonMistakes: { type: string }
                examples: { type: string }
      responses: { '201': { description: OK } }
  /admin/knowledge-points/{id}:
    patch:
      tags: [admin]
      summary: 更新知识点（name/description/commonMistakes/examples/status/parentId）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
...
  /admin/sources/{id}:
    patch:
      tags: [admin]
      summary: 更新来源
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /admin/sources/{id}/sections:
    post:
      tags: [admin]
      summary: 新增章节
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '201': { description: OK } }
...
components:
...
  schemas:
...
    QuestionInput:
      type: object
      required: [type, stem, levelId, subjectId]
      properties:
        type: { type: string, enum: [single_choice, multiple_choice, fill_blank, short_answer] }
        stem: { type: string }
        options:
          type: array
          items:
            type: object
            properties: { id: { type: string }, label: { type: string }, text: { type: string } }
        materialId: { type: string, nullable: true }
        materialTitle: { type: string }
        materialContent: { type: string }
        levelId: { type: string }
        subjectId: { type: string }
        sourceSectionId: { type: string, nullable: true }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        knowledgePointIds: { type: array, items: { type: string } }
        answer:
          nullable: true
          type: object
          description: 省略表示无标准答案（练习时走 AI 判定）
          properties:
            value: { type: object }
            authority: { type: string, enum: [official, human_verified] }
            explanation: { type: string }
    QuestionAdmin:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [draft, in_review, published, retired] }
        hasAnswer: { type: boolean }
        publishedVersionId: { type: string, nullable: true }
        currentVersion: { type: object, nullable: true }
        updatedAt: { type: string }